requirements.txt
bm25s==0.2.13
fitz==0.0.1.dev2
llama_index==0.12.43
nltk==3.9.1
numpy==2.3.1
openai==1.91.0
pandas==1.5.3
pdf2image==1.17.0
PyStemmer==3.0.0
PyStemmer==3.0.0
python-dotenv==1.1.1
Requests==2.32.4
scikit_learn==1.7.0
streamlit==1.30.0